requirements.txt
git+https://github.com/Rapptz/discord.py
python-dotenv
async-timeout==3.0.1
PyNaCl==1.3.0
discordlists.py==1.1.1
asyncio==3.4.3
google-api-python-client==1.10.0